- Don't attempt to output a WinGet Result object if we're going to throw a terminating error.
This situation's complex. The WinGet team's module doesn't throw any error on an operation failure, simply returning the data to the caller for interpretation and throwing. We can't have this as a proper cmdlet/function should throw in the event of an error.

While the previous setup worked from the console's point of view (i.e. you could see the pipeline output even while throwing), you couldn't capture it without proxying it through something like `ForEach-Object` to archive off the pipelined value to a variable.

With this setup, in the module's default state, a terminating error will throw, with the result accessible in the `ErrorRecord`'s `TargetObject` property. For situations such as `-ErrorAction SilentlyContinue`, you'll get the result if `-PassThru` is specified and the resulting `ErrorRecord` will be in the output's `ExtendedErrorCode` property.